Do not process Updates or Finish for Cancel requests
Prior to PR #55 we would not process an update for the cancel request
itself.  While sending progress reports for cancel requests is perfectly
fine from the coordinator perspective it causes tracking mechanics in
lhsmd heartburn of the form:

lhsmd[29378]: 2023/04/27 17:12:39 id:1210 completion failed: no such file or directory

Additionally, calling Finish on a successful cancel request can cause
the archive being canceled to be seen as completed even though it has
not.  We do want to respond to cancel requests that have nothing to
cancel with an error, since that request will have no request with an
identical cookie that's being canceled to respond to the CDT, which
would otherwise clear both out.