Remove unused xattr support for UUID/Hash/URL
Our copytool has never used these, but spent time and effort reading
them, passing them to and fro, and erroring if it didn't find them set.
Get rid of them, which will also allow the hydrator to avoid having to
set them as well.  For historical purposes, the only real use of these
was for snapshot-esque support in historical lemur.  We won't ever use
that, so no danger here.