Disable setting of LOV layout on Restore
Lustre's import llapi function already requires us to set a layout.  At
that time we set the layout equal to the Lustre root to mimic default
layout setting (like a new create would enjoy), but in the future we may
import and set the layout according to a stored metadata attribute in
blob.  In either event, setting the layout again on restore is not useful.