Return ECANCELED for context.Canceled and fix queue counting for cancels
Previously we would return '22' (EINVAL) for 'context canceled', which
was an unfortunate consequence of translating error to a unix error
incompletely.  ECANCELED is a valid error, so return that much more
appropriate rc if error is Canceled.  Because we do this, we can rely on
actions that are successfully canceled to decrement the queue statistic
since the cancel itself doesn't call action.Finish (this confuses the
CDT).  If the cancel is not successful (e.g., cancel for non-existent
action), then Finish is called, so to keep the queue stats correct in
the face of this we allow rpc.go to process the final Update and
decrement the stat.