Support specification of the blob target for archives, removes, and restores by data payload
This change primarily makes functional deletion of files from blob that
no longer exist in Lustre, but also supports the archive of files from
one location in Lustre to a different location in blob, and restore from
a different location in Blob than the target in Lustre.  To support
that, the file id is converted to a string to make specification of the
blob path more straightforward, and the uuid still present in the rpc
payload is re-used to pass this information from the lhsmd to the plugin
(this is how it was previously functioning as well).  Fixed a few
malformed tests in the process.