Kill the parent rather than panicking the plugin
Due to the lack of restart logic in the copytool, panicking just the
plugin can lead to a copytool without a datamover. This will eventually
deregister the agent from the CDT resulting in degraded customer
performance.

Instead of panicking the plugin process, this commit makes the child
send a SIGTERM to the parent when it encounters the "Too busy" error
too many times. SIGTERM forces a shutdown the of the copy tool and all
children, including the process that sent the SIGTERM. Systemd will then
restart the process