Copytool is missing hardlink support on archive
Add support to the mover to archive back to blob all returned paths for
a given FID.  Note that expressly overriding the target blob name was
removed in this commit, as that feature functionality was never used to
our knowledge.

Note that this fix suffers from the deficiency in:
https://jira.whamcloud.com/browse/LU-5746
Which means for maximally-length named files, experiments show that only
15 results will be returned (and therefore archived) by fid2path.  For
more normally-named files seeing north of a hundred links is typical,
so this isn't expected to be a major deficiency for any but truly
esoteric users.